+++ CHANGES.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+### titiler.extensions
+
+* **render**:
+    * change: make `assets` optional in `RenderItem` model
+    * change: add `valid: True|False` key to the `RenderItemWithLinks` model
+    * add: convert old `asset_bix` and `asset_expression` keys to `{asset}|indexes=...&expression=...`
+
 ## 2.0.0a1 (2026-02-11)
 
 ### Misc